add admin component

+++ client/src/components/TheTopBar.vue
@@ -20,8 +20,12 @@ export default {
   name: "TheTopBar",
   methods: {
     redirect() {
-      localStorage.clear();
-      this.$router.push("/Login");
+      if(localStorage.getItem("token")){
+        localStorage.clear();
+      // this.$router.push("/Login");
+      window.location.replace("/Login")
+      }
+      
     },
   },
 };
